scripts/requirements.txt
apache-airflow
Beautifulsoup4
requests
lxml
mysql-connector
pymysql